+++ src/Test/Dictionaries/XmlDictionariesTest.cs
@@ -1,12 +1,12 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
+﻿using System;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Xml;
 using System.Xml.Linq;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Test
+namespace Test.Dictionaries
 {
     [TestClass]
     public class XmlDictionariesTest
